+++ app/layout.tsx
@@ -1,7 +1,8 @@
 import type { Metadata } from 'next'
-import { Cairo } from 'next/font/google'
+import { Cairo,Noto_Kufi_Arabic } from 'next/font/google'
 import '@/styles/globals.css'
 
+const noto = Noto_Kufi_Arabic({ subsets: ['arabic'] })
 const cairo = Cairo({ subsets: ['arabic'] })
 
 export const metadata: Metadata = {
@@ -16,7 +17,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="ar" dir="rtl">
-      <body className={cairo.className}>{children}</body>
+      <body className={noto.className}>{children}</body>
     </html>
   )
 }
