+++ app/page.tsx
@@ -1,109 +1,23 @@
 'use client'
 
 import { useState } from "react"
-import Image from "next/image"
 import { motion } from "framer-motion"
 import { ArrowRight, Menu, X } from "lucide-react"
 import styles from '@/styles/page.module.css'
-const posts = [
-  {
-    title: "مستقبل الذكاء الاصطناعي في تطوير الويب",
-    excerpt: "استكشف كيف يعيد الذكاء الاصطناعي تشكيل مشهد تطوير الويب...",
-    image: "https://plus.unsplash.com/premium_photo-1671209795288-cdbcf73bb53e?q=80&w=1470&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-  },
-  {
-    title: "إتقان React Hooks",
-    excerpt: "تعمق في React Hooks وتعلم كيفية كتابة مكونات React أكثر كفاءة...",
-    image: "https://plus.unsplash.com/premium_photo-1671209795288-cdbcf73bb53e?q=80&w=1470&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-  },
-  {
-    title: "فن التصميم المتجاوب",
-    excerpt: "اكتشف المبادئ والتقنيات وراء إنشاء تصميمات ويب متجاوبة...",
-    image: "https://plus.unsplash.com/premium_photo-1671209795288-cdbcf73bb53e?q=80&w=1470&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-  },
-  {
-    title: "التطورات الحديثة في الأمن السيبراني",
-    excerpt: "تعرف على أحدث الأساليب لحماية مواقع الويب والتطبيقات من الهجمات...",
-    image: "https://plus.unsplash.com/premium_photo-1671209795288-cdbcf73bb53e?q=80&w=1470&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-  },
-  {
-    title: "كيفية بناء تطبيقات ويب قابلة للتوسع",
-    excerpt: "اكتشف المبادئ الأساسية لبناء تطبيقات ويب متطورة وقابلة للتوسع...",
-    image: "https://plus.unsplash.com/premium_photo-1671209795288-cdbcf73bb53e?q=80&w=1470&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-  },
-  {
-    title: "تصميم تجربة مستخدم استثنائية",
-    excerpt: "تعلم كيفية تحسين تجربة المستخدم على مواقعك وتطبيقاتك...",
-    image: "https://plus.unsplash.com/premium_photo-1671209795288-cdbcf73bb53e?q=80&w=1470&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-  }
-]
+import NavBar from "@/components/nav-bar/NavBar"
+import Hero from "@/components/hero/Hero"
+import Post from "@/components/posts/Post"
 
 
 export default function Home() {
-  const [isMenuOpen, setIsMenuOpen] = useState(false)
 
   return (
     <div className={styles.container}>
+      <NavBar />
       <main className={styles.main}>
-        <motion.section 
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ delay: 0.2, duration: 0.8 }}
-          className={styles.hero}
-        >
-          <h2 className={styles.heroTitle}>مرحبًا بك في حديقتي الرقمية</h2>
-          <p className={styles.heroSubtitle}>استكشف الأفكار والرؤى حول التكنولوجيا والحياة</p>
-          <motion.div
-            whileHover={{ scale: 1.05 }}
-            whileTap={{ scale: 0.95 }}
-          >
-            <button className={styles.ctaButton}>
-              ابدأ القراءة
-              <ArrowRight className={styles.arrowIcon} />
-            </button>
-          </motion.div>
-        </motion.section>
-        <motion.section 
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          transition={{ delay: 0.4, duration: 0.8 }}
-          className={styles.posts}
-        >
-          {posts.map((post, i) => (
-            <motion.article
-              key={i}
-              initial={{ opacity: 0, y: 50 }}
-              animate={{ opacity: 1, y: 0 }}
-              transition={{ delay: 0.2 * (i + 1), duration: 0.8 }}
-              className={styles.post}
-            >
-              <motion.div
-                whileHover={{ scale: 1.05 }}
-                transition={{ duration: 0.3 }}
-                className={styles.imageWrapper}
-              >
-                <Image
-                  src={post.image}
-                  alt={post.title}
-                  width={1200}
-                  height={600}
-                  className={styles.postImage}
-                />
-              </motion.div>
-              <div className={styles.postContent}>
-                <h3 className={styles.postTitle}>{post.title}</h3>
-                <p className={styles.postExcerpt}>{post.excerpt}</p>
-                <motion.div
-                  whileHover={{ scale: 1.05 }}
-                  whileTap={{ scale: 0.95 }}
-                >
-                  <button className={styles.readMoreButton}>اقرأ المزيد</button>
-                </motion.div>
-              </div>
-            </motion.article>
-          ))}
-        </motion.section>
-        <motion.section 
+        <Hero />
+        <Post/>
+        <motion.section
           initial={{ opacity: 0, y: 20 }}
           animate={{ opacity: 1, y: 0 }}
           transition={{ delay: 0.8, duration: 0.8 }}
@@ -124,7 +38,7 @@ export default function Home() {
           </form>
         </motion.section>
       </main>
-      <motion.footer 
+      <motion.footer
         initial={{ opacity: 0 }}
         animate={{ opacity: 1 }}
         transition={{ delay: 1, duration: 0.8 }}
