+++ components/posts/styles.module.css
@@ -1,0 +1,68 @@
+.posts {
+  display: grid;
+  gap: 2rem;
+  grid-template-columns: 1fr;
+  /* Single column by default */
+}
+
+.post {
+  background-color: white;
+  border-radius: 8px;
+  overflow: hidden;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  display: flex;
+  flex-direction: column;
+}
+
+.imageWrapper {
+  width: 100%;
+  height: 200px;
+  overflow: hidden;
+}
+
+.postImage {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.postContent {
+  padding: 1.5rem;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+
+}
+
+.postTitle {
+  font-size: 1.5rem;
+  color: #333;
+  margin-bottom: 1rem;
+}
+
+.postExcerpt {
+  color: #666;
+  margin-bottom: 1.5rem;
+}
+
+.readMoreButton {
+  background: linear-gradient(45deg, #20B2AA, #48D1CC);
+  color: white;
+  border: none;
+  padding: 0.5rem 1rem;
+  font-size: 0.9rem;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+@media (min-width: 768px) {
+  .posts {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media (min-width: 1024px) {
+  .posts {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
